+++ mango3-core/src/models/post/post_insert.rs
@@ -57,6 +57,8 @@ impl Post {
                 variables,
                 hashtag_ids,
                 cover_image_blob_id,
+                (SELECT COUNT(*) FROM post_views WHERE post_id = posts.id LIMIT 1) AS "views_count!",
+                (SELECT COUNT(*) FROM post_comments WHERE post_id = posts.id LIMIT 1) AS "comments_count!",
                 published_at,
                 modified_at,
                 NULL::real AS search_rank,