+++ mango3-studio/src/server_functions/posts.rs
@@ -177,9 +177,14 @@ async fn my_post(website_id: &str, id: &str) -> Result<Option<Post>, ServerFnErr
     let core_context = expect_core_context();
     let user = extract_user().await?.unwrap();
 
-    Ok(
-        Post::get_by_id(&core_context, Uuid::try_parse(id)?, Some(&website), Some(&user), None)
-            .await
-            .ok(),
+    Ok(Post::get_by_id(
+        &core_context,
+        Uuid::try_parse(id)?,
+        Some(&website),
+        Some(&user),
+        None,
+        None,
     )
+    .await
+    .ok())
 }