+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - Feature to manage websites themes.
 - Features to search websites and posts.
 - Support for Handlebars on posts content.
+- Features to insert and show post comments.